requirements.txt
mcp
python-dotenv
requests
